+++ sdmx-dl-format-xml/src/main/java/internal/sdmxdl/format/xml/DsdContext.java
@@ -1,20 +1,20 @@
 package internal.sdmxdl.format.xml;
 
 
+import lombok.NonNull;
 import sdmxdl.Codelist;
 import sdmxdl.CodelistRef;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.Optional;
 
 @lombok.Getter
 final class DsdContext {
 
     private final List<Codelist> codelists = new ArrayList<>();
 
-    private final Map<String, String> concepts = new HashMap<>();
+    private final List<Concept> concepts = new ArrayList<>();
 
     private int dimensionCount = 0;
 
@@ -22,11 +22,17 @@ final class DsdContext {
         dimensionCount++;
     }
 
-    public Codelist getCodelist(CodelistRef ref) {
+    public Optional<Codelist> getCodelist(@NonNull CodelistRef ref) {
         return codelists
                 .stream()
                 .filter(ref::containsRef)
-                .findFirst()
-                .orElseGet(() -> Codelist.builder().ref(ref).build());
+                .findFirst();
+    }
+
+    public Optional<Concept> getConcept(@NonNull String id) {
+        return concepts
+                .stream()
+                .filter(concept -> concept.getId().equals(id))
+                .findFirst();
     }
 }